useLoginForm to own file
